refactor(conformance): split probe.rs by concerns (2765 -> 932 mod + 5 submodules)
Audit 2026-05-19 P9 continuation. probe.rs split:

- rest_request.rs  (13 fns: request building + URI label substitution +
                    percent-encoding + form param flattening)
- response.rs       (7 fns: classify_success_expectation + error code
                    extraction + service common errors + classify_response +
                    matches_declared_error + truncate)
- protocols.rs      (4 fns: probe_query / probe_json / probe_rest /
                    sigv4_auth_header)
- service_detection.rs (2 fns: service_protocol / is_aws_query_compatible_service)
- followup.rs       (2 fns: reader_required_inputs / run_round_trip_followup)
- mod.rs            (probe_variant entry points + Protocol enum + tests)
